ief_paving/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ief"
version = "0.1.0"
description = "Industry Entity Finder (paving pilot: TX, MI, CO)"
authors = [{name="Your Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "aiohttp>=3.9.5",
  "cdx_toolkit>=0.9.8",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "trafilatura>=1.9.0",
  "aiolimiter>=1.1.0",
  "requests>=2.31.0",
  "pandas>=2.2.2",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "rapidfuzz>=3.9.6",
  "phonenumbers>=8.13.40",
  "tldextract>=5.1.2",
  "duckdb>=1.0.0",
  "pyyaml>=6.0.2",
]